+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+
+
 ## 0.8.1
 
 - Improved most string parser builders. The generated code is fast again
 - Minor improvements in parser builders `Alpha0`, `Alpha1`, `Alphanumeric0`, `Alphanumeric1`, `Digit0`, `Digit1`, `HexDigit0`, `HexDigit1`
+- Removed parser builders `Crlf`, `LineEnding` in favor of using parser builder `Tags` which generates code more efficiently than the original code for these removed builders
 
 ## 0.8.0